+++ packages/pwa-kit-mcp/CHANGELOG.md
@@ -5,3 +5,4 @@
 - Add `create_app_guidelines` MCP tool. [#2599](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2599)
 - Add `run_site_test` MCP tool [#2645](https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2645):
 - Add `create_sample_storefront_page` MCP tool [#2899] (https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2899)
+- fix `logMCPMessage` and add a test for this debugging util. [#2989] (https://github.com/SalesforceCommerceCloud/pwa-kit/pull/2989)